Move the DynamoDB gateway into a sub-module
