+++ src/gateways/dynamodb/awsSession.go
@@ -1,4 +1,4 @@
-package gateways
+package dynamodb
 
 import "github.com/aws/aws-sdk-go/aws/session"
